1_model_deployment/pyproject.toml
[tool.poetry]
name = "model-deployment"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "model_deployment"}]

[tool.poetry.dependencies]
python = "3.10.11"
fastapi = "^0.96.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
scikit-learn = "^1.2.2"
joblib = "^1.2.0"
numpy = "^1.24.3"
google-cloud-storage = "^2.9.0"
loguru = "^0.7.0"



[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
